Add in TTF support in ... again
So this time I was a bit more successful with adding in TTF support.
Earlier the bitmaps would not render right in the renderer, due to some
bugs in the renderer itself. Now that those are fixed, the fonts get
rendered in the right orientation and do resemble somewhat the font that
they need to resemble.

However the fonts do not render perfecly, this due to some format
problems, which will be fixed in later commits.