Dont round the x and y pos for getting the frame_buffer pixel
This results in a massive performance boost as these std::round calls
are quite performance intensive.

Also in this commit, have i removed the function call for setting the
shmaddr. I've tried making the function call inline, but this did't gave
as good of a result as doing it this way.