Add more simple image space clipping
This method is much and much easier than the one I was trying to
implement. And is also much more scalable with new attributes, like when
we want to do perspective texture mapping. Also this method is not a
whole lot slower. Not in the big picture atleas, because most polygons
are already culled or are within the view frustrum.